check perms on buttons

+++ src/main/java/org/cascadebot/cascadebot/commands/music/PlayingCommand.java
@@ -16,6 +16,7 @@ import org.cascadebot.cascadebot.data.objects.Flag;
 import org.cascadebot.cascadebot.messaging.MessagingObjects;
 import org.cascadebot.cascadebot.music.CascadePlayer;
 import org.cascadebot.cascadebot.permissions.CascadePermission;
+import org.cascadebot.cascadebot.permissions.PermissionsManager;
 import org.cascadebot.cascadebot.utils.FormatUtils;
 import org.cascadebot.cascadebot.utils.buttons.Button;
 import org.cascadebot.cascadebot.utils.buttons.ButtonGroup;
@@ -25,26 +26,38 @@ import java.util.Set;
 public class PlayingCommand implements ICommandMain {
 
     private Button.UnicodeButton playButton = new Button.UnicodeButton(UnicodeConstants.PLAY, (runner, channel, message) -> {
-        handlePlayPause(GuildDataManager.getGuildData(channel.getGuild().getIdLong()).getButtonsCache().get(channel.getIdLong()).get(message.getIdLong()), message);
+        if (CascadeBot.INS.getPermissionsManager().isAuthorised(CascadeBot.INS.getCommandManager().getCommandByDefault("resume"), GuildDataManager.getGuildData(channel.getGuild().getIdLong()), runner)) {
+            handlePlayPause(GuildDataManager.getGuildData(channel.getGuild().getIdLong()).getButtonsCache().get(channel.getIdLong()).get(message.getIdLong()), message);
+        }
     });
 
     private Button.UnicodeButton pauseButton = new Button.UnicodeButton(UnicodeConstants.PAUSE, (runner, channel, message) -> {
-        handlePlayPause(GuildDataManager.getGuildData(channel.getGuild().getIdLong()).getButtonsCache().get(channel.getIdLong()).get(message.getIdLong()), message);
+        if (CascadeBot.INS.getPermissionsManager().isAuthorised(CascadeBot.INS.getCommandManager().getCommandByDefault("pause"), GuildDataManager.getGuildData(channel.getGuild().getIdLong()), runner)) {
+            handlePlayPause(GuildDataManager.getGuildData(channel.getGuild().getIdLong()).getButtonsCache().get(channel.getIdLong()).get(message.getIdLong()), message);
+        }
     });
 
     private Button.UnicodeButton repeat = new Button.UnicodeButton(UnicodeConstants.REPEAT, (runner, channel, message) -> {
-        ButtonGroup buttonGroup = GuildDataManager.getGuildData(channel.getGuild().getIdLong()).getButtonsCache().get(channel.getIdLong()).get(message.getIdLong());
-        handleRepeat(buttonGroup, CascadePlayer.LoopMode.PLAYLIST, message);
+
+        if (CascadeBot.INS.getPermissionsManager().isAuthorised(CascadeBot.INS.getCommandManager().getCommandByDefault("loop"), GuildDataManager.getGuildData(channel.getGuild().getIdLong()), runner)) {
+            ButtonGroup buttonGroup = GuildDataManager.getGuildData(channel.getGuild().getIdLong()).getButtonsCache().get(channel.getIdLong()).get(message.getIdLong());
+            handleRepeat(buttonGroup, CascadePlayer.LoopMode.PLAYLIST, message);
+        }
     });
 
     private Button.UnicodeButton repeatOne = new Button.UnicodeButton(UnicodeConstants.REPEAT_ONCE, (runner, channel, message) -> {
-        ButtonGroup buttonGroup = GuildDataManager.getGuildData(channel.getGuild().getIdLong()).getButtonsCache().get(channel.getIdLong()).get(message.getIdLong());
-        handleRepeat(buttonGroup, CascadePlayer.LoopMode.SONG, message);
+        if (CascadeBot.INS.getPermissionsManager().isAuthorised(CascadeBot.INS.getCommandManager().getCommandByDefault("loop"), GuildDataManager.getGuildData(channel.getGuild().getIdLong()), runner)) {
+            ButtonGroup buttonGroup = GuildDataManager.getGuildData(channel.getGuild().getIdLong()).getButtonsCache().get(channel.getIdLong()).get(message.getIdLong());
+            handleRepeat(buttonGroup, CascadePlayer.LoopMode.SONG, message);
+        }
     });
 
     private Button.EmoteButton noRepeat = new Button.EmoteButton(Config.INS.getGlobalEmotes().get("norepeat"), (runner, channel, message) -> {
-        ButtonGroup buttonGroup = GuildDataManager.getGuildData(channel.getGuild().getIdLong()).getButtonsCache().get(channel.getIdLong()).get(message.getIdLong());
-        handleRepeat(buttonGroup, CascadePlayer.LoopMode.DISABLED, message);
+
+        if (CascadeBot.INS.getPermissionsManager().isAuthorised(CascadeBot.INS.getCommandManager().getCommandByDefault("loop"), GuildDataManager.getGuildData(channel.getGuild().getIdLong()), runner)) {
+            ButtonGroup buttonGroup = GuildDataManager.getGuildData(channel.getGuild().getIdLong()).getButtonsCache().get(channel.getIdLong()).get(message.getIdLong());
+            handleRepeat(buttonGroup, CascadePlayer.LoopMode.DISABLED, message);
+        }
     });
 
     @Override
@@ -57,44 +70,52 @@ public class PlayingCommand implements ICommandMain {
             ButtonGroup buttonGroup = new ButtonGroup(sender.getUser().getIdLong(), context.getChannel().getIdLong(), context.getGuild().getIdLong());
             if (context.getData().isFlagEnabled(Flag.MUSIC_SERVICES)) {
                 buttonGroup.addButton(new Button.UnicodeButton(UnicodeConstants.VOLUME_DOWN, (runner, channel, message) -> {
-                    int volume = context.getMusicPlayer().getPlayer().getVolume();
-                    volume -= 10;
-                    if (volume <= 0) {
-                        volume = 0;
+                    if (context.hasPermission(runner, "volume")) {
+                        int volume = context.getMusicPlayer().getPlayer().getVolume();
+                        volume -= 10;
+                        if (volume <= 0) {
+                            volume = 0;
+                        }
+                        context.getMusicPlayer().getPlayer().setVolume(volume);
+                        message.editMessage(getSongEmbed(player, context.getGuild().getIdLong())).queue();
                     }
-                    context.getMusicPlayer().getPlayer().setVolume(volume);
-                    message.editMessage(getSongEmbed(player, context.getGuild().getIdLong())).queue();
                 }));
                 buttonGroup.addButton(new Button.UnicodeButton(UnicodeConstants.VOLUME_UP, (runner, channel, message) -> {
-                    int volume = context.getMusicPlayer().getPlayer().getVolume();
-                    volume += 10;
-                    if (volume >= 100) {
-                        volume = 100;
+                    if (context.hasPermission(runner, "volume")) {
+                        int volume = context.getMusicPlayer().getPlayer().getVolume();
+                        volume += 10;
+                        if (volume >= 100) {
+                            volume = 100;
+                        }
+                        context.getMusicPlayer().getPlayer().setVolume(volume);
+                        message.editMessage(getSongEmbed(player, context.getGuild().getIdLong())).queue();
                     }
-                    context.getMusicPlayer().getPlayer().setVolume(volume);
-                    message.editMessage(getSongEmbed(player, context.getGuild().getIdLong())).queue();
                 }));
             }
 
             buttonGroup.addButton(new Button.UnicodeButton(UnicodeConstants.STOP, (runner, channel, message) -> {
-                context.getMusicPlayer().stop();
-                message.delete().queue();
+                if (context.hasPermission(runner, "stop")) {
+                    context.getMusicPlayer().stop();
+                    message.delete().queue();
+                }
             }));
             buttonGroup.addButton(new Button.UnicodeButton(UnicodeConstants.FAST_FORWARD, (runner, channel, message) -> {
-                CascadeBot.INS.getCommandManager().getCommandByDefault("skip").onCommand(runner, new CommandContext(
-                        CascadeBot.INS.getClient(),
-                        context.getChannel(),
-                        message,
-                        context.getGuild(),
-                        context.getData(),
-                        new String[0],
-                        runner,
-                        "skip",
-                        false
-                ));
-                message.editMessage(getSongEmbed(player, context.getGuild().getIdLong())).queue();
-                if (player.getPlayer().getPlayingTrack() == null) {
-                    message.clearReactions().queue();
+                if (context.hasPermission(runner, "skip")) {
+                    CascadeBot.INS.getCommandManager().getCommandByDefault("skip").onCommand(runner, new CommandContext(
+                            CascadeBot.INS.getClient(),
+                            context.getChannel(),
+                            message,
+                            context.getGuild(),
+                            context.getData(),
+                            new String[0],
+                            runner,
+                            "skip",
+                            false
+                    ));
+                    message.editMessage(getSongEmbed(player, context.getGuild().getIdLong())).queue();
+                    if (player.getPlayer().getPlayingTrack() == null) {
+                        message.clearReactions().queue();
+                    }
                 }
             }));
